+++ src/Controller/IuForm/IuFormAbstractController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller\User;
+namespace App\Controller\IuForm;
 
 use App\Controller\Traits\CreatorByCreatorIdTrait;
 use App\Repository\CreatorRepository;
