+++ src/Controller/IuForm/IuFormUtils/StartData.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller\User\IuFormUtils;
+namespace App\Controller\IuForm\IuFormUtils;
 
 // No validation here. If you play with the scripts and something goes wrong, blame yourself.
 class StartData
